requirements.txt
aiogram
psycopg2-binary
requests
pillow
redis
